slit core js
